Added dimensionality of the "screen" as designed
Set a container object in main to depict the appropriate boundaries of the pseudo-screen that the users will see. Also positioned the 'buttons' for transitioning between pages in the appropriate arrangement (top, middle, bottom). Please ignore ugly color palates. Also added a settings button but it has some conflicts with routed objects so I commented it out. Will fix soon.